Glasser Multi-Modal Cortical Parcellation (HCP)

Brain atlas for the Glasser cortical parcellation with 361 regions based on the Human Connectome Project multi-modal parcellation. Contains 2D polygon geometry for ggseg::geom_brain().

Usage

glasser()

Value

A ggseg.formats::ggseg_atlas object (cortical).

References

Glasser MF, Coalson TS, Robinson EC, Hacker CD, Harwell J, Yacoub E, Ugurbil K, Andersson J, Beckmann CF, Jenkinson M, Smith SM, Van Essen DC (2016). A multi-modal parcellation of human cerebral cortex. Nature, 536(7615):171-178. doi:10.1038/nature18933

Examples

glasser()
#> 
#> ── glasser ggseg atlas ─────────────────────────────────────────────────────────
#> Type: cortical
#> Regions: 180
#> Hemispheres: left, right
#> Views: inferior, lateral, medial, superior
#> Palette: ✔
#> Rendering: ✔ ggseg
#> ✔ ggseg3d (vertices)
#> ────────────────────────────────────────────────────────────────────────────────
#> # A tibble: 360 × 3
#>     hemi  region label          
#>     <chr> <chr>  <chr>          
#>   1 left  V1     lh_L_V1_ROI    
#>   2 left  MST    lh_L_MST_ROI   
#>   3 left  V6     lh_L_V6_ROI    
#>   4 left  V2     lh_L_V2_ROI    
#>   5 left  V3     lh_L_V3_ROI    
#>   6 left  V4     lh_L_V4_ROI    
#>   7 left  V8     lh_L_V8_ROI    
#>   8 left  4      lh_L_4_ROI     
#>   9 left  3b     lh_L_3b_ROI    
#>  10 left  FEF    lh_L_FEF_ROI   
#>  11 left  PEF    lh_L_PEF_ROI   
#>  12 left  55b    lh_L_55b_ROI   
#>  13 left  V3A    lh_L_V3A_ROI   
#>  14 left  RSC    lh_L_RSC_ROI   
#>  15 left  POS2   lh_L_POS2_ROI  
#>  16 left  V7     lh_L_V7_ROI    
#>  17 left  IPS1   lh_L_IPS1_ROI  
#>  18 left  FFC    lh_L_FFC_ROI   
#>  19 left  V3B    lh_L_V3B_ROI   
#>  20 left  LO1    lh_L_LO1_ROI   
#>  21 left  LO2    lh_L_LO2_ROI   
#>  22 left  PIT    lh_L_PIT_ROI   
#>  23 left  MT     lh_L_MT_ROI    
#>  24 left  A1     lh_L_A1_ROI    
#>  25 left  PSL    lh_L_PSL_ROI   
#>  26 left  SFL    lh_L_SFL_ROI   
#>  27 left  PCV    lh_L_PCV_ROI   
#>  28 left  STV    lh_L_STV_ROI   
#>  29 left  7Pm    lh_L_7Pm_ROI   
#>  30 left  7m     lh_L_7m_ROI    
#>  31 left  POS1   lh_L_POS1_ROI  
#>  32 left  23d    lh_L_23d_ROI   
#>  33 left  v23ab  lh_L_v23ab_ROI 
#>  34 left  d23ab  lh_L_d23ab_ROI 
#>  35 left  31pv   lh_L_31pv_ROI  
#>  36 left  5m     lh_L_5m_ROI    
#>  37 left  5mv    lh_L_5mv_ROI   
#>  38 left  23c    lh_L_23c_ROI   
#>  39 left  5L     lh_L_5L_ROI    
#>  40 left  24dd   lh_L_24dd_ROI  
#>  41 left  24dv   lh_L_24dv_ROI  
#>  42 left  7AL    lh_L_7AL_ROI   
#>  43 left  SCEF   lh_L_SCEF_ROI  
#>  44 left  6ma    lh_L_6ma_ROI   
#>  45 left  7Am    lh_L_7Am_ROI   
#>  46 left  7PL    lh_L_7PL_ROI   
#>  47 left  7PC    lh_L_7PC_ROI   
#>  48 left  LIPv   lh_L_LIPv_ROI  
#>  49 left  VIP    lh_L_VIP_ROI   
#>  50 left  MIP    lh_L_MIP_ROI   
#>  51 left  1      lh_L_1_ROI     
#>  52 left  2      lh_L_2_ROI     
#>  53 left  3a     lh_L_3a_ROI    
#>  54 left  6d     lh_L_6d_ROI    
#>  55 left  6mp    lh_L_6mp_ROI   
#>  56 left  6v     lh_L_6v_ROI    
#>  57 left  p24pr  lh_L_p24pr_ROI 
#>  58 left  33pr   lh_L_33pr_ROI  
#>  59 left  a24pr  lh_L_a24pr_ROI 
#>  60 left  p32pr  lh_L_p32pr_ROI 
#>  61 left  a24    lh_L_a24_ROI   
#>  62 left  d32    lh_L_d32_ROI   
#>  63 left  8BM    lh_L_8BM_ROI   
#>  64 left  p32    lh_L_p32_ROI   
#>  65 left  10r    lh_L_10r_ROI   
#>  66 left  47m    lh_L_47m_ROI   
#>  67 left  8Av    lh_L_8Av_ROI   
#>  68 left  8Ad    lh_L_8Ad_ROI   
#>  69 left  9m     lh_L_9m_ROI    
#>  70 left  8BL    lh_L_8BL_ROI   
#>  71 left  9p     lh_L_9p_ROI    
#>  72 left  10d    lh_L_10d_ROI   
#>  73 left  8C     lh_L_8C_ROI    
#>  74 left  44     lh_L_44_ROI    
#>  75 left  45     lh_L_45_ROI    
#>  76 left  47l    lh_L_47l_ROI   
#>  77 left  a47r   lh_L_a47r_ROI  
#>  78 left  6r     lh_L_6r_ROI    
#>  79 left  IFJa   lh_L_IFJa_ROI  
#>  80 left  IFJp   lh_L_IFJp_ROI  
#>  81 left  IFSp   lh_L_IFSp_ROI  
#>  82 left  IFSa   lh_L_IFSa_ROI  
#>  83 left  p9-46v lh_L_p9-46v_ROI
#>  84 left  46     lh_L_46_ROI    
#>  85 left  a9-46v lh_L_a9-46v_ROI
#>  86 left  9-46d  lh_L_9-46d_ROI 
#>  87 left  9a     lh_L_9a_ROI    
#>  88 left  10v    lh_L_10v_ROI   
#>  89 left  a10p   lh_L_a10p_ROI  
#>  90 left  10pp   lh_L_10pp_ROI  
#>  91 left  11l    lh_L_11l_ROI   
#>  92 left  13l    lh_L_13l_ROI   
#>  93 left  OFC    lh_L_OFC_ROI   
#>  94 left  47s    lh_L_47s_ROI   
#>  95 left  LIPd   lh_L_LIPd_ROI  
#>  96 left  6a     lh_L_6a_ROI    
#>  97 left  i6-8   lh_L_i6-8_ROI  
#>  98 left  s6-8   lh_L_s6-8_ROI  
#>  99 left  43     lh_L_43_ROI    
#> 100 left  OP4    lh_L_OP4_ROI   
#> 101 left  OP1    lh_L_OP1_ROI   
#> 102 left  OP2-3  lh_L_OP2-3_ROI 
#> 103 left  52     lh_L_52_ROI    
#> 104 left  RI     lh_L_RI_ROI    
#> 105 left  PFcm   lh_L_PFcm_ROI  
#> 106 left  PoI2   lh_L_PoI2_ROI  
#> 107 left  TA2    lh_L_TA2_ROI   
#> 108 left  FOP4   lh_L_FOP4_ROI  
#> 109 left  MI     lh_L_MI_ROI    
#> 110 left  Pir    lh_L_Pir_ROI   
#> 111 left  AVI    lh_L_AVI_ROI   
#> 112 left  AAIC   lh_L_AAIC_ROI  
#> 113 left  FOP1   lh_L_FOP1_ROI  
#> 114 left  FOP3   lh_L_FOP3_ROI  
#> 115 left  FOP2   lh_L_FOP2_ROI  
#> 116 left  PFt    lh_L_PFt_ROI   
#> 117 left  AIP    lh_L_AIP_ROI   
#> 118 left  EC     lh_L_EC_ROI    
#> 119 left  PreS   lh_L_PreS_ROI  
#> 120 left  H      lh_L_H_ROI     
#> 121 left  ProS   lh_L_ProS_ROI  
#> 122 left  PeEc   lh_L_PeEc_ROI  
#> 123 left  STGa   lh_L_STGa_ROI  
#> 124 left  PBelt  lh_L_PBelt_ROI 
#> 125 left  A5     lh_L_A5_ROI    
#> 126 left  PHA1   lh_L_PHA1_ROI  
#> 127 left  PHA3   lh_L_PHA3_ROI  
#> 128 left  STSda  lh_L_STSda_ROI 
#> 129 left  STSdp  lh_L_STSdp_ROI 
#> 130 left  STSvp  lh_L_STSvp_ROI 
#> 131 left  TGd    lh_L_TGd_ROI   
#> 132 left  TE1a   lh_L_TE1a_ROI  
#> 133 left  TE1p   lh_L_TE1p_ROI  
#> 134 left  TE2a   lh_L_TE2a_ROI  
#> 135 left  TF     lh_L_TF_ROI    
#> 136 left  TE2p   lh_L_TE2p_ROI  
#> 137 left  PHT    lh_L_PHT_ROI   
#> 138 left  PH     lh_L_PH_ROI    
#> 139 left  TPOJ1  lh_L_TPOJ1_ROI 
#> 140 left  TPOJ2  lh_L_TPOJ2_ROI 
#> 141 left  TPOJ3  lh_L_TPOJ3_ROI 
#> 142 left  DVT    lh_L_DVT_ROI   
#> 143 left  PGp    lh_L_PGp_ROI   
#> 144 left  IP2    lh_L_IP2_ROI   
#> 145 left  IP1    lh_L_IP1_ROI   
#> 146 left  IP0    lh_L_IP0_ROI   
#> 147 left  PFop   lh_L_PFop_ROI  
#> 148 left  PF     lh_L_PF_ROI    
#> 149 left  PFm    lh_L_PFm_ROI   
#> 150 left  PGi    lh_L_PGi_ROI   
#> 151 left  PGs    lh_L_PGs_ROI   
#> 152 left  V6A    lh_L_V6A_ROI   
#> 153 left  VMV1   lh_L_VMV1_ROI  
#> 154 left  VMV3   lh_L_VMV3_ROI  
#> 155 left  PHA2   lh_L_PHA2_ROI  
#> 156 left  V4t    lh_L_V4t_ROI   
#> 157 left  FST    lh_L_FST_ROI   
#> 158 left  V3CD   lh_L_V3CD_ROI  
#> 159 left  LO3    lh_L_LO3_ROI   
#> 160 left  VMV2   lh_L_VMV2_ROI  
#> 161 left  31pd   lh_L_31pd_ROI  
#> 162 left  31a    lh_L_31a_ROI   
#> 163 left  VVC    lh_L_VVC_ROI   
#> 164 left  25     lh_L_25_ROI    
#> 165 left  s32    lh_L_s32_ROI   
#> 166 left  pOFC   lh_L_pOFC_ROI  
#> 167 left  PoI1   lh_L_PoI1_ROI  
#> 168 left  Ig     lh_L_Ig_ROI    
#> 169 left  FOP5   lh_L_FOP5_ROI  
#> 170 left  p10p   lh_L_p10p_ROI  
#> 171 left  p47r   lh_L_p47r_ROI  
#> 172 left  TGv    lh_L_TGv_ROI   
#> 173 left  MBelt  lh_L_MBelt_ROI 
#> 174 left  LBelt  lh_L_LBelt_ROI 
#> 175 left  A4     lh_L_A4_ROI    
#> 176 left  STSva  lh_L_STSva_ROI 
#> 177 left  TE1m   lh_L_TE1m_ROI  
#> 178 left  PI     lh_L_PI_ROI    
#> 179 left  a32pr  lh_L_a32pr_ROI 
#> 180 left  p24    lh_L_p24_ROI   
#> 181 right V1     rh_R_V1_ROI    
#> 182 right MST    rh_R_MST_ROI   
#> 183 right V6     rh_R_V6_ROI    
#> 184 right V2     rh_R_V2_ROI    
#> 185 right V3     rh_R_V3_ROI    
#> 186 right V4     rh_R_V4_ROI    
#> 187 right V8     rh_R_V8_ROI    
#> 188 right 4      rh_R_4_ROI     
#> 189 right 3b     rh_R_3b_ROI    
#> 190 right FEF    rh_R_FEF_ROI   
#> 191 right PEF    rh_R_PEF_ROI   
#> 192 right 55b    rh_R_55b_ROI   
#> 193 right V3A    rh_R_V3A_ROI   
#> 194 right RSC    rh_R_RSC_ROI   
#> 195 right POS2   rh_R_POS2_ROI  
#> 196 right V7     rh_R_V7_ROI    
#> 197 right IPS1   rh_R_IPS1_ROI  
#> 198 right FFC    rh_R_FFC_ROI   
#> 199 right V3B    rh_R_V3B_ROI   
#> 200 right LO1    rh_R_LO1_ROI   
#> 201 right LO2    rh_R_LO2_ROI   
#> 202 right PIT    rh_R_PIT_ROI   
#> 203 right MT     rh_R_MT_ROI    
#> 204 right A1     rh_R_A1_ROI    
#> 205 right PSL    rh_R_PSL_ROI   
#> 206 right SFL    rh_R_SFL_ROI   
#> 207 right PCV    rh_R_PCV_ROI   
#> 208 right STV    rh_R_STV_ROI   
#> 209 right 7Pm    rh_R_7Pm_ROI   
#> 210 right 7m     rh_R_7m_ROI    
#> 211 right POS1   rh_R_POS1_ROI  
#> 212 right 23d    rh_R_23d_ROI   
#> 213 right v23ab  rh_R_v23ab_ROI 
#> 214 right d23ab  rh_R_d23ab_ROI 
#> 215 right 31pv   rh_R_31pv_ROI  
#> 216 right 5m     rh_R_5m_ROI    
#> 217 right 5mv    rh_R_5mv_ROI   
#> 218 right 23c    rh_R_23c_ROI   
#> 219 right 5L     rh_R_5L_ROI    
#> 220 right 24dd   rh_R_24dd_ROI  
#> 221 right 24dv   rh_R_24dv_ROI  
#> 222 right 7AL    rh_R_7AL_ROI   
#> 223 right SCEF   rh_R_SCEF_ROI  
#> 224 right 6ma    rh_R_6ma_ROI   
#> 225 right 7Am    rh_R_7Am_ROI   
#> 226 right 7PL    rh_R_7PL_ROI   
#> 227 right 7PC    rh_R_7PC_ROI   
#> 228 right LIPv   rh_R_LIPv_ROI  
#> 229 right VIP    rh_R_VIP_ROI   
#> 230 right MIP    rh_R_MIP_ROI   
#> 231 right 1      rh_R_1_ROI     
#> 232 right 2      rh_R_2_ROI     
#> 233 right 3a     rh_R_3a_ROI    
#> 234 right 6d     rh_R_6d_ROI    
#> 235 right 6mp    rh_R_6mp_ROI   
#> 236 right 6v     rh_R_6v_ROI    
#> 237 right p24pr  rh_R_p24pr_ROI 
#> 238 right 33pr   rh_R_33pr_ROI  
#> 239 right a24pr  rh_R_a24pr_ROI 
#> 240 right p32pr  rh_R_p32pr_ROI 
#> 241 right a24    rh_R_a24_ROI   
#> 242 right d32    rh_R_d32_ROI   
#> 243 right 8BM    rh_R_8BM_ROI   
#> 244 right p32    rh_R_p32_ROI   
#> 245 right 10r    rh_R_10r_ROI   
#> 246 right 47m    rh_R_47m_ROI   
#> 247 right 8Av    rh_R_8Av_ROI   
#> 248 right 8Ad    rh_R_8Ad_ROI   
#> 249 right 9m     rh_R_9m_ROI    
#> 250 right 8BL    rh_R_8BL_ROI   
#> 251 right 9p     rh_R_9p_ROI    
#> 252 right 10d    rh_R_10d_ROI   
#> 253 right 8C     rh_R_8C_ROI    
#> 254 right 44     rh_R_44_ROI    
#> 255 right 45     rh_R_45_ROI    
#> 256 right 47l    rh_R_47l_ROI   
#> 257 right a47r   rh_R_a47r_ROI  
#> 258 right 6r     rh_R_6r_ROI    
#> 259 right IFJa   rh_R_IFJa_ROI  
#> 260 right IFJp   rh_R_IFJp_ROI  
#> 261 right IFSp   rh_R_IFSp_ROI  
#> 262 right IFSa   rh_R_IFSa_ROI  
#> 263 right p9-46v rh_R_p9-46v_ROI
#> 264 right 46     rh_R_46_ROI    
#> 265 right a9-46v rh_R_a9-46v_ROI
#> 266 right 9-46d  rh_R_9-46d_ROI 
#> 267 right 9a     rh_R_9a_ROI    
#> 268 right 10v    rh_R_10v_ROI   
#> 269 right a10p   rh_R_a10p_ROI  
#> 270 right 10pp   rh_R_10pp_ROI  
#> 271 right 11l    rh_R_11l_ROI   
#> 272 right 13l    rh_R_13l_ROI   
#> 273 right OFC    rh_R_OFC_ROI   
#> 274 right 47s    rh_R_47s_ROI   
#> 275 right LIPd   rh_R_LIPd_ROI  
#> 276 right 6a     rh_R_6a_ROI    
#> 277 right i6-8   rh_R_i6-8_ROI  
#> 278 right s6-8   rh_R_s6-8_ROI  
#> 279 right 43     rh_R_43_ROI    
#> 280 right OP4    rh_R_OP4_ROI   
#> 281 right OP1    rh_R_OP1_ROI   
#> 282 right OP2-3  rh_R_OP2-3_ROI 
#> 283 right 52     rh_R_52_ROI    
#> 284 right RI     rh_R_RI_ROI    
#> 285 right PFcm   rh_R_PFcm_ROI  
#> 286 right PoI2   rh_R_PoI2_ROI  
#> 287 right TA2    rh_R_TA2_ROI   
#> 288 right FOP4   rh_R_FOP4_ROI  
#> 289 right MI     rh_R_MI_ROI    
#> 290 right Pir    rh_R_Pir_ROI   
#> 291 right AVI    rh_R_AVI_ROI   
#> 292 right AAIC   rh_R_AAIC_ROI  
#> 293 right FOP1   rh_R_FOP1_ROI  
#> 294 right FOP3   rh_R_FOP3_ROI  
#> 295 right FOP2   rh_R_FOP2_ROI  
#> 296 right PFt    rh_R_PFt_ROI   
#> 297 right AIP    rh_R_AIP_ROI   
#> 298 right EC     rh_R_EC_ROI    
#> 299 right PreS   rh_R_PreS_ROI  
#> 300 right H      rh_R_H_ROI     
#> 301 right ProS   rh_R_ProS_ROI  
#> 302 right PeEc   rh_R_PeEc_ROI  
#> 303 right STGa   rh_R_STGa_ROI  
#> 304 right PBelt  rh_R_PBelt_ROI 
#> 305 right A5     rh_R_A5_ROI    
#> 306 right PHA1   rh_R_PHA1_ROI  
#> 307 right PHA3   rh_R_PHA3_ROI  
#> 308 right STSda  rh_R_STSda_ROI 
#> 309 right STSdp  rh_R_STSdp_ROI 
#> 310 right STSvp  rh_R_STSvp_ROI 
#> 311 right TGd    rh_R_TGd_ROI   
#> 312 right TE1a   rh_R_TE1a_ROI  
#> 313 right TE1p   rh_R_TE1p_ROI  
#> 314 right TE2a   rh_R_TE2a_ROI  
#> 315 right TF     rh_R_TF_ROI    
#> 316 right TE2p   rh_R_TE2p_ROI  
#> 317 right PHT    rh_R_PHT_ROI   
#> 318 right PH     rh_R_PH_ROI    
#> 319 right TPOJ1  rh_R_TPOJ1_ROI 
#> 320 right TPOJ2  rh_R_TPOJ2_ROI 
#> 321 right TPOJ3  rh_R_TPOJ3_ROI 
#> 322 right DVT    rh_R_DVT_ROI   
#> 323 right PGp    rh_R_PGp_ROI   
#> 324 right IP2    rh_R_IP2_ROI   
#> 325 right IP1    rh_R_IP1_ROI   
#> 326 right IP0    rh_R_IP0_ROI   
#> 327 right PFop   rh_R_PFop_ROI  
#> 328 right PF     rh_R_PF_ROI    
#> 329 right PFm    rh_R_PFm_ROI   
#> 330 right PGi    rh_R_PGi_ROI   
#> 331 right PGs    rh_R_PGs_ROI   
#> 332 right V6A    rh_R_V6A_ROI   
#> 333 right VMV1   rh_R_VMV1_ROI  
#> 334 right VMV3   rh_R_VMV3_ROI  
#> 335 right PHA2   rh_R_PHA2_ROI  
#> 336 right V4t    rh_R_V4t_ROI   
#> 337 right FST    rh_R_FST_ROI   
#> 338 right V3CD   rh_R_V3CD_ROI  
#> 339 right LO3    rh_R_LO3_ROI   
#> 340 right VMV2   rh_R_VMV2_ROI  
#> 341 right 31pd   rh_R_31pd_ROI  
#> 342 right 31a    rh_R_31a_ROI   
#> 343 right VVC    rh_R_VVC_ROI   
#> 344 right 25     rh_R_25_ROI    
#> 345 right s32    rh_R_s32_ROI   
#> 346 right pOFC   rh_R_pOFC_ROI  
#> 347 right PoI1   rh_R_PoI1_ROI  
#> 348 right Ig     rh_R_Ig_ROI    
#> 349 right FOP5   rh_R_FOP5_ROI  
#> 350 right p10p   rh_R_p10p_ROI  
#> 351 right p47r   rh_R_p47r_ROI  
#> 352 right TGv    rh_R_TGv_ROI   
#> 353 right MBelt  rh_R_MBelt_ROI 
#> 354 right LBelt  rh_R_LBelt_ROI 
#> 355 right A4     rh_R_A4_ROI    
#> 356 right STSva  rh_R_STSva_ROI 
#> 357 right TE1m   rh_R_TE1m_ROI  
#> 358 right PI     rh_R_PI_ROI    
#> 359 right a32pr  rh_R_a32pr_ROI 
#> 360 right p24    rh_R_p24_ROI